requirements.txt
streamlit
numpy
pillow
cryptography
pyswarm
scipy
